refactor: simplify build and packaging of layout package
