+++ packages/layout/bs-column.js
@@ -2,11 +2,11 @@
 import { LitElement, html } from 'lit-element';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
-import { BsColumnExtraSmallCss } from './bs-column-xs.css.js';
-import { BsColumnSmallCss } from './bs-column-sm.css.js';
-import { BsColumnMediumCss } from './bs-column-md.css.js';
-import { BsColumnLargeCss } from './bs-column-lg.css.js';
-import { BsColumnExtraLargeCss } from './bs-column-xl.css.js';
+import { BsColumnExtraSmallCss } from './css/bs-column-xs.css.js';
+import { BsColumnSmallCss } from './css/bs-column-sm.css.js';
+import { BsColumnMediumCss } from './css/bs-column-md.css.js';
+import { BsColumnLargeCss } from './css/bs-column-lg.css.js';
+import { BsColumnExtraLargeCss } from './css/bs-column-xl.css.js';
 
 export class BsColumn extends LitElement {
