+++ packages/layout/css/index.js
@@ -1,8 +1,4 @@
-
-export { BsRow } from './bs-row.js';
 export { BsRowCss } from './bs-row.css.js';
-export { BsColumn } from './bs-column.js';
-export { BsContainer } from './bs-container.js';
 export { BsContainerCss } from './bs-container.css.js';
 export { BsColumnExtraSmallCss } from './bs-column-xs.css.js';
 export { BsColumnSmallCss } from './bs-column-sm.css.js';
